extract 'import project' button to standalone component
